requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0

# NLP
sentence-transformers>=2.2.0
transformers>=4.30.0
scikit-learn>=1.3.0
torch>=2.0.0

# Data storage
pyarrow>=12.0.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
pytest>=7.4.0

# SEC EDGAR specific
python-dotenv>=1.0.0
sec-edgar-downloader>=5.0.0
